reorganise oav devices and rename snapshot to MJPG
but keep derivative devices named snapshot
